test: add makefiles test command (converted from script)
